Added shadow cubemap blend.
I changed the materials all to black, removed the windows and put a solid mesh at the fuselage / flight deck - I'm not sure if this is right as I don't have enough specific DHC6 aircraft knowledge.

The prop discs and blades are hidden in the blend as I don't think it's right to show these.